libmport: DMA Query enhancement
When querying the status of a DMA channel, it is useful to know
which device the DMA channel is interacting with.  The DMA query
service is enhanced to return the DestID along with the channel
number.  The destID is in the most significant 2 bytes of the
parameter, and the channel number is in the least significant 2
bytes.  A destID value of 0xFFFF means that the channel is idle.