TSI721 UMD: Only start engine 6
Change the mask used to call umd_queu_config_multi to only initialize
engine 6.  A mask of 0xFF will initialize engine 7, which is hard coded
in the driver to be used by the kernel for all remote register access.
In other words, after initialization the kernel cannot access remote
registers but it will still try...